backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
google-cloud-vision>=3.7.0
pytest>=8.0.0
requests>=2.31.0

